Move tests for should_autorenew
